Local Relay, config validation and custom image optimization
Local Replay implementation, which removes the requirement for configuring subnets and IPs of the workers in plex, as well as the use of the FORCE_HTTPS setting when Plex is using Secure Connections = Required.
Docker images have been optimized for reuse.
Grafana sample has been improved.
Reorg of documentation
Adding Kubernetes documentation